+++ MagicAPI/src/main/java/com/elmakers/mine/bukkit/api/arena/Arena.java
@@ -1,0 +1,5 @@
+package com.elmakers.mine.bukkit.api.arena;
+
+public interface Arena {
+    String getKey();
+}