+++ CompatibilityLib/v1_19_3/src/main/java/com/elmakers/mine/bukkit/utility/platform/v1_19_3/ItemUtils.java
@@ -6,7 +6,7 @@ import java.util.List;
 import java.util.logging.Logger;
 
 import org.bukkit.Material;
-import org.bukkit.craftbukkit.v1_19_R1.inventory.CraftItemStack;
+import org.bukkit.craftbukkit.v1_19_R2.inventory.CraftItemStack;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.inventory.meta.ItemMeta;