+++ Magic/CHANGELOG.md
@@ -3,6 +3,8 @@
 
 # 10.8.5
 
+- Add format option to individual currencies for custom number formatting per currency type
+
 # 10.8.4
 
  - Revert change from 10.8.2: A spell is no longer allowed to damage or combust an entity it is not allowed to target